various fixes and optimizations

+++ test.sh
@@ -16,9 +16,10 @@ LOG_DIR=./log
 #MAP=./test/maps/long_coord_test.txt
 #MAP=./test/maps/headon_symmetric_test.txt
 #MAP=./test/maps/dont_split_test.txt
+MAP=./test/maps/blocking_test.txt
 #MAP=$KIT_DIR/maps/apocalyptic.txt
 #MAP=$KIT_DIR/maps/empty-room.txt
-MAP=$KIT_DIR/maps/huge-room.txt
+#MAP=$KIT_DIR/maps/huge-room.txt
 #MAP=$KIT_DIR/maps/playground.txt
 
 #MYBOT_FILE=./lib/main.rb
@@ -32,12 +33,15 @@ MYBOT_CMD="ruby $MYBOT_FILE --debug"
 RIVAL_FILE=$KIT_DIR/example_bots/WallHugger.jar
 RIVAL_CMD="$JAVA_CMD -jar $RIVAL_FILE"
 
-rm $LOG_DIR/fight.log
-for i in $KIT_DIR/maps/*.txt; do
-    $JAVA_CMD -jar $KIT_DIR/engine/Tron.jar $i "$MYBOT_CMD" "$RIVAL_CMD" 0 1 2>&1 >> $LOG_DIR/fight.log
-done
+if [ "$1" = "-a" ]; then
+    rm $LOG_DIR/fight.log
+    for i in $KIT_DIR/maps/*.txt; do
+       $JAVA_CMD -jar $KIT_DIR/engine/Tron.jar $i "$MYBOT_CMD" "$RIVAL_CMD" 0 1 2>&1 >> $LOG_DIR/fight.log
+    done
+else
+    $JAVA_CMD -jar $KIT_DIR/engine/Tron.jar $MAP "$MYBOT_CMD" "$RIVAL_CMD" 0 1 | tee $LOG_DIR/fight.log
+fi
 
-#$JAVA_CMD -jar $KIT_DIR/engine/Tron.jar $MAP "$MYBOT_CMD" "$RIVAL_CMD" 0 1 | tee $LOG_DIR/fight.log
 
 echo "`basename $MYBOT_FILE .rb` | `basename $RIVAL_FILE .jar` | `basename $MAP` | `tail -n 1 $LOG_DIR/fight.log`" >> $LOG_DIR/stat.txt 1.5
